Create playGame function and alter playSingleRound
Return values were added to playSingleRound to
add some logic to the newly created function.
It returns 0 if the user won the round, 1 if
the computer won the round and 2 if there was
a tie.

playGame function makes the whole logic of the
game, which consists of turning the rounds,
telling who won each round and the final game.
It tells if a tie happened too.